+++ NEWS.md
@@ -1,3 +1,15 @@
+# meantables 0.1.2
+
+* Added the `mean_format()` function.
+
+# meantables 0.1.1
+
+* Changed the way that `n_miss` and `n` are calculated. Now correct when there are missing values. Built unit test to check for this.
+
+* Changed the `x` argument of `mean_table()` to `.x`. This prevents an error when passing a variable named `x` to `mean_table()`.
+
+* Added standard deviation to `mean_table()` results.
+
 # meantables 0.1.0
 
 * First release on CRAN